[DEL/#72] DateText.kt 삭제 후 해당 코드 DiaryWriteScreen에 추가
